More prep of shader files.
